+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/files/AudioFile.kt
@@ -2,6 +2,8 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.files
 
 import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.Performerable
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.FileId
+import com.github.insanusmokrassar.TelegramBotAPI.types.FileUniqueId
+import com.github.insanusmokrassar.TelegramBotAPI.types.fileUniqueIdField
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.abstracts.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -10,6 +12,8 @@ import kotlinx.serialization.Serializable
 data class AudioFile(
     @SerialName(fileIdField)
     override val fileId: FileId,
+    @SerialName(fileUniqueIdField)
+    override val fileUniqueId: FileUniqueId,
     override val duration: Long? = null,
     override val performer: String? = null,
     override val title: String? = null,